Update Fluree test client, example for data constraint policy
- Distinct functions for testing read and writes
- Additional write function for constraint testing
- Separate data for each function
- Write data with random UUID to keep inserting anew
- Proper namespaces and prefixes via JSON-LD `@context`

If you try to submit a Project with the same ID twice but with a
different name, the constraint on the name will be violated. Fluree
returns a HTTP 422, but if the proxy doesn't handle that, it will be a
500.

The write test will keep writing the same entity with different random
IDs. You will see that entity on subsequent runs, not in the same run
that both the read and write occur.

The violation write will fail, as the violation is triggered with the
same data as the policy write, which is commented out (run it
uncommented once).

A reminder that if you have errors starting Fluree with Docker, change
the data volume name and create the ledger anew.